+++ src/components/notification/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./component";
+export { useNotification } from "./component/context";
+export * from "./types";