+++ src/components/skeleton/components/index.tsx
@@ -1,0 +1,4 @@
+export { default as SkeletonText } from "./skeletonText";
+export { default as SkeletonBlock } from "./skeletonBlock";
+export { default as SkeletonAvatar } from "./skeletonAvatar";
+export { default as SkeletonButton } from "./skeletonButton";